Final commit

+++ app/Http/Controllers/GroupController.php
@@ -57,17 +57,45 @@ class GroupController extends Controller
 
     public function deleteUserFromGroup($id)
     {
-        $user = User::find(Auth::user()->id);
-        $group = Group::where('id', $id)->first();
-        $user->groups()->detach($group);
-        return redirect('groups');
+        if (Auth::user()->groups->where('id', $id)->first() && Auth::user()->groups->where('id', $id)->first()->pivot->is_moderator) {
+            return back();
+        } else {
+            $user = User::find(Auth::user()->id);
+            $group = Group::where('id', $id)->first();
+            $user->groups()->detach($group);
+            return redirect('groups');
+        }
     }
 
     public function delete($id)
     {
         $group = Group::where('id', $id)->first();
+        Post::where('group_id', $id)->delete();
         $group->users()->detach();
         $group->delete();
         return redirect('groups');
+    }
+
+    public function transferRights(Request $request, $id)
+    {
+        $newModerator = $request->new_moderator;
+        $group = Group::find($id);
+        $currentModerator = $group->users()->wherePivot('is_moderator', true)->first();
+        $currentModerator->groups()->updateExistingPivot($id, ['is_moderator' => false]);
+        $newMod = User::find($newModerator);
+        $newMod->groups()->updateExistingPivot($id, ['is_moderator' => true]);
+        return redirect()->back();
+    }
+    
+    public function searchGroup(Request $request){
+        $query = $request->post('query');
+
+        if (empty($query)) {
+            $groups = Group::all();
+            return redirect('groups');
+        } else {
+            $groups = Group::where('name', 'like', "%{$query}%")->get();
+        }
+        return view('groups', compact('groups', 'query'));
     }
 }
